% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{AdicionarBotones}
\alias{AdicionarBotones}
\title{Agregar columnas con botones HTML interactivos}
\usage{
AdicionarBotones(...)
}
\arguments{
  \item{tabla}{`data.frame` al que se agregan los botones.}
  \item{botones}{Vector de caracteres con los nombres de los botones.}
}
\value{`data.frame` con nuevas columnas HTML.}
\description{
Agregar columnas con botones HTML interactivos
}
