% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{Espacios}
\alias{Espacios}
\title{Insertar espacios no separables HTML}
\usage{
Espacios(...)
}
\arguments{
  \item{n}{Numero de entidades `&nbsp;` a insertar.}
}
\value{Objeto HTML de Shiny.}
\description{
Insertar espacios no separables HTML
}
\examples{
Espacios(3)
}
