% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{Obligatorio}
\alias{Obligatorio}
\title{Marcar texto como campo obligatorio con asterisco rojo}
\usage{
Obligatorio(...)
}
\arguments{
  \item{s}{Texto de la etiqueta.}
}
\value{Objeto HTML con asterisco rojo al final.}
\description{
Marcar texto como campo obligatorio con asterisco rojo
}
\examples{
Obligatorio("Nombre del cliente")
}
