% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{RedondearMultiplo}
\alias{RedondearMultiplo}
\title{Redondear al multiplo mas cercano}
\usage{
RedondearMultiplo(...)
}
\arguments{
  \item{x}{Numero o vector numerico.}
  \item{multiple}{Multiplo objetivo.}
}
\value{Numero redondeado al multiplo mas cercano.}
\description{
Redondear al multiplo mas cercano
}
\examples{
RedondearMultiplo(17, 5)
RedondearMultiplo(c(13, 17, 22), 5)
}
