% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{Saltos}
\alias{Saltos}
\title{Generar saltos de linea HTML}
\usage{
Saltos(...)
}
\arguments{
  \item{n}{Numero de saltos de linea `<br>` a insertar.}
}
\value{Objeto HTML de Shiny.}
\description{
Generar saltos de linea HTML
}
\examples{
Saltos(2)
}
