% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{Unicos}
\alias{Unicos}
\title{Valores unicos ordenados de un vector}
\usage{
Unicos(...)
}
\arguments{
  \item{x}{Vector de cualquier tipo.}
}
\value{Vector con valores unicos en orden ascendente.}
\description{
Valores unicos ordenados de un vector
}
\examples{
Unicos(c("b", "a", "a", "c"))
}
